Flushed

+++ src/main/java/com/buuz135/industrial/block/misc/tile/EnchantmentApplicatorTile.java
@@ -40,7 +40,7 @@ import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraftforge.fluids.FluidStack;
+import io.github.fabricators_of_create.porting_lib.util.FluidStack;
 import net.minecraftforge.fluids.capability.CapabilityFluidHandler;
 import net.minecraftforge.fluids.capability.IFluidHandler;
 import org.apache.commons.lang3.tuple.Pair;
@@ -123,7 +123,7 @@ public class EnchantmentApplicatorTile extends IndustrialProcessingTile<Enchantm
     }
 
     @Override
-    protected int getTickPower() {
+    protected long getTickPower() {
         return EnchantmentApplicatorConfig.powerPerTick;
     }
