+++ src/main/java/com/buuz135/industrial/block/resourceproduction/tile/FluidSievingMachineTile.java
@@ -115,7 +115,7 @@ public class FluidSievingMachineTile extends IndustrialProcessingTile<FluidSievi
     }
 
     @Override
-    protected int getTickPower() {
+    protected long getTickPower() {
         return FluidSievingMachineConfig.powerPerTick;
     }
