+++ src/main/java/com/buuz135/industrial/config/machine/resourceproduction/FermentationStationConfig.java
@@ -30,7 +30,7 @@ import com.hrznstudio.titanium.annotation.config.ConfigVal;
 public class FermentationStationConfig {
 
     @ConfigVal(comment = "Amount of Power Consumed per Tick - Default: [40FE]")
-    public static int powerPerTick = 40;
+    public static long powerPerTick = 40;
 
     @ConfigVal(comment = "Max Stored Power [FE] - Default: [10000 FE]")
     public static int maxStoredPower = 10000;
